requirements.txt
django==3.0
dj-database-url==0.5.0
psycopg2==2.8.4
python-decouple==3.3
Unipath==1.1
gunicorn
django-heroku
graphene-django>=2.0
pytest-django
